backend/requirements.txt
Django==6.0.3
psycopg[binary]>=3.1,<4.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
drf-spectacular>=0.27,<1.0
